+++ client/src/app/main/modules/main-content/modules/larger-screen-content/larger-screen-content.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { LargeScreenContentComponent } from './components/large-sreen-content/large-screen-content.component';
 import { CustomTranslateModule } from '../../../../../shared/modules/translate/custom-translate.module';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatToolbarModule } from '@angular/material/toolbar';
@@ -15,12 +14,12 @@ import { GridModule } from '../../../../../shared/modules/grid/grid.module';
 
 @NgModule({
   declarations: [
-    LargeScreenContentComponent,
     TitleBarComponent,
     FooterComponent
   ],
   exports: [
-    LargeScreenContentComponent
+    TitleBarComponent,
+    FooterComponent
   ],
   imports: [
     CommonModule,