+++ client/src/app/shared/modules/button/button.module.ts
@@ -3,33 +3,36 @@ import {CommonModule} from '@angular/common';
 import {ButtonIconComponent} from './components/button-icon/button-icon.component';
 import {IconModule} from '../icon/icon.module';
 import { MatButtonModule } from '@angular/material/button';
-import { MatDialogModule } from '@angular/material/dialog';
 import {ButtonComponent} from './components/button/button.component';
 import {ButtonRowComponent} from './components/button-row/button-row.component';
 import {CustomTranslateModule} from '../translate/custom-translate.module';
 import {GridModule} from '../grid/grid.module';
 import { ButtonIconActiveStateComponent } from './components/button-icon-active-state/button-icon-active-state.component';
+import { MatDialogModule } from '@angular/material/dialog';
+import { ButtonNoElevationComponent } from './components/button-no-elevation/button-no-elevation.component';
 
 @NgModule({
   declarations: [
     ButtonComponent,
     ButtonRowComponent,
     ButtonIconComponent,
-    ButtonIconActiveStateComponent
+    ButtonIconActiveStateComponent,
+    ButtonNoElevationComponent
   ],
   exports: [
     ButtonComponent,
     ButtonRowComponent,
     ButtonIconComponent,
-    ButtonIconActiveStateComponent
+    ButtonIconActiveStateComponent,
+    ButtonNoElevationComponent
   ],
   imports: [
     CommonModule,
     IconModule,
     MatButtonModule,
     CustomTranslateModule,
-    MatDialogModule,
-    GridModule
+    GridModule,
+    MatDialogModule
   ]
 })
 export class ButtonModule { }