+++ client/src/app/shared/modules/user-action/components/user-actions/user-actions.component.ts
@@ -11,7 +11,7 @@ import { AuthService } from 'src/app/shared/modules/auth/services/auth/auth.serv
 })
 export class UserActionsComponent implements OnInit {
 
-  @ViewChild('userSettingsDropDown', {static: true})
+  @ViewChild('userSettingsDropDown')
   public userSettingsDropDown: MatSelect;
 
   @Input()