+++ e2e/src/app.po.ts
@@ -33,7 +33,7 @@ export class AppPage {
   }
 
   public getWhiteDropDownList(): ElementFinder {
-    return this.helper.findById('white-user');
+    return this.helper.selectLastElementWithId('white-user');
   }
 
   public getWhiteDropDownField(): ElementFinder {
@@ -41,23 +41,11 @@ export class AppPage {
   }
 
   public getUser(id: string): ElementFinder {
-    return this.helper.findById(id);
-  }
-
-  public getProtractorUser(): ElementFinder {
-    return this.helper.findById('protractor');
+    return this.helper.selectLastElementWithId(id);
   }
 
   public getWinnerDropDownList(): ElementFinder {
-    return this.helper.findById('winner-user');
-  }
-
-  public getWhiteOption(): ElementFinder {
-    return this.helper.findById('White');
-  }
-
-  public getBlackOption(): ElementFinder {
-    return this.helper.findById('Black');
+    return this.helper.selectLastElementWithId('winner-user');
   }
 
   public getWinnerDropDownField(): ElementFinder {
@@ -65,23 +53,19 @@ export class AppPage {
   }
 
   public getBlackDropDownList(): ElementFinder {
-    return this.helper.findById('black-user');
+    return this.helper.selectLastElementWithId('black-user');
   }
 
   public getBlackDropDownField(): ElementFinder {
     return this.helper.findByXPath('//*[@id="black-user"]/div/div[1]/span');
   }
 
-  public getProtractorUser1(): ElementFinder {
-    return this.helper.findById('protractor1');
-  }
-
   public getSaveButton(): ElementFinder {
-    return this.helper.findById('save');
+    return this.helper.selectLastElementWithId('save');
   }
 
   public getTableWhiteUserForLastAddedGame(): ElementFinder {
-    return this.helper.findById('table-protractor-0');
+    return this.helper.selectLastElementWithId('table-protractor-0');
   }
 
   public getTableWinnerUserForLastAddedGame(winner: string): ElementFinder {
@@ -89,7 +73,7 @@ export class AppPage {
   }
 
   public getTableBlackUserForLastAddedGame(): ElementFinder {
-    return this.helper.findById('table-protractor1-0');
+    return this.helper.selectLastElementWithId('table-protractor1-0');
   }
 
   public getGamesView(): ElementFinder {