+++ e2e/src/helper.ts
@@ -88,24 +88,21 @@ export class Helper {
 
   public addGame(page: any, browser: any, game: any): void {
     page.getWhiteDropDownList().click();
-
+    
     page.getUser(game.white).click();
 
-    browser.waitForAngular();
     expect(page.getWhiteDropDownField().getText()).toEqual(game.white);
 
     page.getBlackDropDownList().click();
 
     page.getUser(game.black).click();
     
-    browser.waitForAngular();
     expect(page.getBlackDropDownField().getText()).toEqual(game.black);
-
+    
     page.getWinnerDropDownList().click();
 
     page.getUser(game.winner).click();
-
-    browser.waitForAngular();
+    
     expect(page.getWinnerDropDownField().getText()).toEqual(game.winner);
 
     page.getSaveButton().click();